Refactored FeedItemStats so it can be used used for both Articles and Notes
ref https://linear.app/ghost/issue/AP-559/add-post-stats-and-buttons-to-articles-in-drawer-view
